deepen: HashAssign owns allocation invariant (254 lines)
Inline alphabet (ALPH/HASH_RE) + hasher (xxh32/contentChecksum/initHasher)
+ pure (lineHashesPure/mapStableHashes/canon, HASH_SPACE=62^3) into
hash-assign.ts. Previously split across 4 files with interface≈implementation;
now one deep module owns the probe-stride allocation invariant.

Shallow files alphabet/hasher/pure/hash become thin shims re-exporting
from hash-assign (backward compat); hash.ts keeps persistence wrapper
importing pure APIs from hash-assign (no circular). Callers already
migrated to hash-assign in phase 2a; index is thin barrel via deep seams.

Fix HL_PREFIX_MINUS_RE regex escaping (was {{ANCHOR_LEN}}). Typecheck +
674 tests pass (was 2 failing on -   │ stripping).
