+++ src/hashline/hasher.ts
@@ -1,33 +1,2 @@
-import xxhash from "xxhash-wasm";
-
-export type Hasher = {
-	h32(input: string, seed?: number): number;
-	h64ToString(input: string, seed?: bigint): string;
-};
-
-let hasher: Hasher | null = null;
-
-export function getH(): Hasher {
-	if (hasher) return hasher;
-	throw new Error("xxhash-wasm hasher not initialized; await initHasher() before calling hashline APIs.");
-}
-
-const hasherP: Promise<Hasher> = xxhash().then((h) => {
-	hasher = h as unknown as Hasher;
-	return hasher;
-}).catch((err: unknown) => {
-	console.error("xxhash-wasm initialization failed:", err);
-	throw err;
-});
-
-export function initHasher(): Promise<Hasher> {
-	return hasherP;
-}
-
-export function xxh32(input: string, seed = 0): number {
-	return getH().h32(input, seed) >>> 0;
-}
-
-export function contentChecksum(content: string): string {
-	return getH().h64ToString(content);
-}
+// @deprecated — private to HashAssign seam. Use `from "./hash-assign.js"` instead.
+export * from "./hash-assign.js";
