+++ src/mutation.ts
@@ -32,24 +32,40 @@ import type { FsSandboxController } from "./sandbox.js";
 import { execPipeline, snapshotIdFor } from "./edit-pipeline.js";
 import type { PipelineResult } from "./edit-pipeline.js";
 import {
-  runFileEdits,
-  resolveMissingPath,
-  persistUndoAndWrite,
-  enforceNoopLoop,
-  collectRemovedHashes,
-  countLineChanges,
+ runFileEdits,
+ resolveMissingPath,
+ persistUndoAndWrite,
+ enforceNoopLoop,
+ collectRemovedHashes,
+ countLineChanges,
 } from "./edit-engine.js";
 import type { FileEditResult, PreparedItem } from "./edit-engine.js";
-import { buildMetrics, buildNoop, buildChanged, buildBatchResult } from "./edit-response.js";
+import {
+ buildMetrics,
+ buildNoop,
+ buildChanged,
+ buildBatchResult,
+} from "./edit-response.js";
 import type { RMeta, BatchSection } from "./edit-response.js";
 import { genDiff, restoreEndings, toLF, stripBOM } from "./edit-diff.js";
 import { computeDrift, scanDrift } from "./drift.js";
-import { trackNoopPayload, clearNoopLoop, noopPayloadKey } from "./noop-guard.js";
+import {
+ trackNoopPayload,
+ clearNoopLoop,
+ noopPayloadKey,
+} from "./noop-guard.js";
 
 // Re-exports for callers that need the underlying types/helpers
 export { execPipeline, snapshotIdFor };
 export type { PipelineResult };
-export { runFileEdits, resolveMissingPath, persistUndoAndWrite, enforceNoopLoop, collectRemovedHashes, countLineChanges };
+export {
+ runFileEdits,
+ resolveMissingPath,
+ persistUndoAndWrite,
+ enforceNoopLoop,
+ collectRemovedHashes,
+ countLineChanges,
+};
 export type { FileEditResult, PreparedItem };
 export { buildMetrics, buildNoop, buildChanged, buildBatchResult };
 export type { RMeta, BatchSection };
@@ -61,50 +77,55 @@ export { trackNoopPayload, clearNoopLoop, noopPayloadKey };
 
 /** Apply a single edit — owns read→normalize→loadServed→applyOne→stableRehash→drift. */
 export async function applySingle(
-  io: FileIO,
-  params: EditParams,
-  cwd: string,
-  opts?: { sessionKey?: string; signal?: AbortSignal; store?: HashStore; noPersist?: boolean },
+ io: FileIO,
+ params: EditParams,
+ cwd: string,
+ opts?: {
+  sessionKey?: string;
+  signal?: AbortSignal;
+  store?: HashStore;
+  noPersist?: boolean;
+ },
 ): Promise<PipelineResult> {
-  return execPipeline(io, params, cwd, opts);
+ return execPipeline(io, params, cwd, opts);
 }
 
 /** Apply a per-file sequence (batch's group) — owns the loop + unionRange + counters. */
 export async function applySequence(
-  io: FileIO,
-  items: PreparedItem[],
-  ctx: { sessionKey: string; signal?: AbortSignal },
+ io: FileIO,
+ items: PreparedItem[],
+ ctx: { sessionKey: string; signal?: AbortSignal },
 ): Promise<FileEditResult> {
-  return runFileEdits(io, items, ctx);
+ return runFileEdits(io, items, ctx);
 }
 
 /** Commit the transaction — owns persist-undo → write → restore. */
 export async function commit(opts: {
-  io: FileIO;
-  files: Array<{
-    absolutePath: string;
-    displayPath: string;
-    originalNormalized: string;
-    bom: string;
-    originalEnding: import("./edit-diff.js").LineEnding;
-    originalHashes: string[];
-    result: string;
-  }>;
-  exec: ToolExecution;
-  sandbox: FsSandboxController;
-  sandboxPolicy: SandboxExecutionPolicy | undefined;
-  signal?: AbortSignal;
-  undoUnavailableMessage: (displayPath: string) => string;
-  restoreUnwrittenUndos?: boolean;
+ io: FileIO;
+ files: Array<{
+  absolutePath: string;
+  displayPath: string;
+  originalNormalized: string;
+  bom: string;
+  originalEnding: import("./edit-diff.js").LineEnding;
+  originalHashes: string[];
+  result: string;
+ }>;
+ exec: ToolExecution;
+ sandbox: FsSandboxController;
+ sandboxPolicy: SandboxExecutionPolicy | undefined;
+ signal?: AbortSignal;
+ undoUnavailableMessage: (displayPath: string) => string;
+ restoreUnwrittenUndos?: boolean;
 }): Promise<void> {
-  return persistUndoAndWrite({
-    io: opts.io,
-    files: opts.files,
-    exec: opts.exec,
-    sandbox: opts.sandbox,
-    sandboxPolicy: opts.sandboxPolicy,
-    signal: opts.signal,
-    undoUnavailableMessage: opts.undoUnavailableMessage,
-    restoreUnwrittenUndos: opts.restoreUnwrittenUndos,
-  });
+ return persistUndoAndWrite({
+  io: opts.io,
+  files: opts.files,
+  exec: opts.exec,
+  sandbox: opts.sandbox,
+  sandboxPolicy: opts.sandboxPolicy,
+  signal: opts.signal,
+  undoUnavailableMessage: opts.undoUnavailableMessage,
+  restoreUnwrittenUndos: opts.restoreUnwrittenUndos,
+ });
 }
