+++ src/tool-undo.ts
@@ -12,7 +12,8 @@ import { cntDiff, splitLines } from "./utils.js";
 import { assertUndoRequest } from "./contract.js";
 import { normalizeRequest as normReq } from "./contract.js";
 import { upsertSnapshotFor } from "./hash-store.js";
-import { contentChecksum, lineHashes } from "./hashline/hash-assign.js";
+import { contentChecksum } from "./hashline/hash-assign.js";
+import { lineHashes } from "./hashline/hash.js";
 import { changedRange } from "./hashline/anchor-pipeline.js";
 import { getUndo, clearUndo } from "./undo-edit.js";
 import { recordServedTruncated } from "./session-view.js";
